Functionality added thanks to SecSan. SecSan wrote: I needed some additional functionality like a way of only adding new vertices/edges if they aren't already present in the graph.
The GetOrCreate queries allow to only create a new vertex or edge if it doesn't exist already. This assumes that edges are unique between two vertices and a given edge label. For vertices, an existing vertex is searched based on an indexed property.

UpdateEdge is equivalent to UpdateVertex and allows to update edge properties.

This pull request also includes small fixes, e.g., to use parametrized scripts where they weren't used before.